configurable auto-remove for wallet, pool

+++ von_agent/nodepool.py
@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from indy import pool
+from indy import pool, IndyError
+from indy.error import ErrorCode
 
 import json
 import logging
@@ -25,20 +26,28 @@ class NodePool:
     Class encapsulating indy-sdk node pool.
     """
 
-    def __init__(self, name: str, genesis_txn_path: str) -> None:
+    def __init__(self, name: str, genesis_txn_path: str, cfg: dict = None) -> None:
         """
         Initializer for node pool. Does not open the pool, only retains input parameters.
 
         :param name: name of the pool
         :param genesis_txn_path: path to genesis transaction file
+        :param cfg: configuration, None for default;
+            i.e., {
+                'auto_remove': bool (default False), whether to remove serialized indy configuration data on close
+            }
         """
 
         logger = logging.getLogger(__name__)
-        logger.debug('NodePool.__init__: >>> name: {}, genesis_txn_path: {}'.format(name, genesis_txn_path))
+        logger.debug('NodePool.__init__: >>> name: {}, genesis_txn_path: {}, cfg: {}'.format(
+            name,
+            genesis_txn_path,
+            cfg))
 
         self._name = name
         self._genesis_txn_path = genesis_txn_path
         self._handle = None
+        self._cfg = cfg or {}
 
         logger.debug('NodePool.__init__: <<<')
 
@@ -99,7 +108,15 @@ class NodePool:
         logger = logging.getLogger(__name__)
         logger.debug('NodePool.open: >>>')
 
-        await pool.create_pool_ledger_config(self.name, json.dumps({'genesis_txn': str(self.genesis_txn_path)}))
+        try:
+            await pool.create_pool_ledger_config(self.name, json.dumps({'genesis_txn': str(self.genesis_txn_path)}))
+        except IndyError as e:
+            if e.error_code == ErrorCode.PoolLedgerConfigAlreadyExistsError:
+                logger.info('Pool ledger config for {} already exists'.format(self.name))
+            else:
+                logger.error('NodePool cannot create ledger configuration: indy error code {}'.format(e.error_code))
+                raise e
+
         self._handle = await pool.open_pool_ledger(self.name, None)
 
         logger.debug('NodePool.open: <<<')
@@ -132,7 +149,8 @@ class NodePool:
         logger.debug('NodePool.close: >>>')
 
         await pool.close_pool_ledger(self.handle)
-        await pool.delete_pool_ledger_config(self.name)
+        if self._cfg and 'auto_remove' in self._cfg and self._cfg['auto_remove']:
+            await pool.delete_pool_ledger_config(self.name)
 
         logger.debug('NodePool.close: <<<')
